feat: install vim8
This was kind of annoying and probably isn't the best way to do it. Vim8
doesn't appear to be available without installing a new repository, the
command for which the base image doesn't include (`add-apt-repository`).

`vi` still points to the old version, to that needs to uninstalled and
replaced (via alias here).